- A feedback system to raise or lower the limit on the max allowed and max manual_transaction depending on the feedback from the SUPPORT role.
- GET Transaction History
- GET Transaction History By Number
- Custom Feedback Serializer.

+++ Anti-Fraud System/task/src/antifraud/presentation/DTO/transaction/TransactionRequestDTO.java
@@ -8,7 +8,7 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
 
-public record TransactionRequestDTO(@NotNull long amount,
+public record TransactionRequestDTO(@NotNull int amount,
                                     @ValidIP String ip,
                                     @ValidCardNumber String number,
                                     Region region,
